[Review] Apply suggestions
