+++ Sources/App/Extensions/String+ContentExtensions.swift
@@ -15,15 +15,3 @@ extension String {
     }
     
 }
-
-extension Optional where Wrapped == String {
-    
-    func isNullOrBlank() -> Bool {
-        return self.wrapped == nil || self.unsafelyUnwrapped.isBlank()
-    }
-    
-    func isNotNullNorBlank() -> Bool {
-        return !self.isNullOrBlank()
-    }
-    
-}
